[REFACTOR]: 사용자인증토큰 메소드 사용

+++ src/main/java/KGUcapstone/OutDecision/domain/comments/service/CommentsService.java
@@ -10,6 +10,8 @@ import KGUcapstone.OutDecision.domain.post.repository.PostRepository;
 import KGUcapstone.OutDecision.domain.user.domain.Member;
 import KGUcapstone.OutDecision.domain.user.repository.MemberRepository;
 import KGUcapstone.OutDecision.domain.user.service.FindMemberService;
+import KGUcapstone.OutDecision.global.error.exception.handler.MemberHandler;
+import KGUcapstone.OutDecision.global.error.status.ErrorStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -41,15 +43,23 @@ public class CommentsService {
 
             return comment;
         } else {
-            throw new RuntimeException("User not found");
+            throw new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND);
         }
     }
 
 
     @Transactional
     public void delete(Long postsId, Long commentsId) {
+        Optional<Member> memberOptional = findMemberService.findLoginMember();
+        Long memberId;
+        // 로그인 체크
+        if(memberOptional.isPresent()) memberId = memberOptional.get().getId();
+        else throw new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND);
+
         Comments comments = commentsRepository.findByPostIdAndId(postsId, commentsId).orElseThrow(() ->
                 new IllegalArgumentException("해당 댓글이 존재하지 않습니다. commentsId=" + commentsId));
+
+        if (!memberId.equals(comments.getMember().getId())) throw new IllegalArgumentException("Bad Request");
 
         commentsRepository.delete(comments);
     }
